refractor and report

+++ recordsystem/RecordManager.cpp
@@ -1,57 +1,61 @@
 #include "RecordManager.h"
 
-void RecordManager::newTable(string tableName){
+RecordManager::RecordManager(){
+    this->tablePool.clear();
+}
+
+shared_ptr<TableInfo> RecordManager::newTable(string tableName){
+    if (this->tablePool.find(tableName) != this->tablePool.end()){
+        cout << "Table " << tableName << " already exists!\n";
+        return nullptr;
+    }
     ofstream tInfoOut(tableName+".tbinfo");
     tInfoOut.close();
     shared_ptr<TableInfo> tif = make_shared<TableInfo>(tableName, true);
-    this->tinfo = tif;
-    this->infoFileName = tableName+".tbinfo";
-    this->dataFileName = tableName+".tbdata";
+    this->tablePool[tableName] = tif;
+    return tif;
 }
 
-void RecordManager::newTable(string tableName, vector<pair<string, varTypes>> cols, vector<int> sizes){
-    newTable(tableName);
+shared_ptr<TableInfo> RecordManager::newTable(string tableName, vector<pair<string, varTypes>> cols, vector<int> sizes){
+    if (this->tablePool.find(tableName) != this->tablePool.end()){
+        cout << "Table " << tableName << " already exists!\n";
+        return nullptr;
+    }
+    shared_ptr<TableInfo> tif = newTable(tableName);
     int ind = 0;
     for (unsigned int i=0;i<cols.size();i++){
         if ((cols[i].second == varTypes::VARCHAR_TYPE) || (cols[i].second == varTypes::CHAR_TYPE)){
-            this->tinfo->addNewColumn(cols[i].first, cols[i].second, sizes[ind]);
+            tif->addNewColumn(cols[i].first, cols[i].second, sizes[ind]);
             ind++;
         } else {
-            this->tinfo->addNewColumn(cols[i].first, cols[i].second, DataOperands::getTypeSize(cols[i].second));
+            tif->addNewColumn(cols[i].first, cols[i].second, DataOperands::getTypeSize(cols[i].second));
         }
     }
-    openDataFile();
+    openDataFile(tif);
+    this->tablePool[tableName] = tif;
+    return tif;
 }
 
-void RecordManager::openTable(string tableName){
+shared_ptr<TableInfo> RecordManager::openTable(string tableName){
+    if (this->tablePool.find(tableName) != this->tablePool.end()){
+        return this->tablePool[tableName];
+    }
     shared_ptr<TableInfo> tif = make_shared<TableInfo>(tableName);
-    this->tinfo = tif;
-    this->infoFileName = tableName+".tbinfo";
-    this->dataFileName = tableName+".tbdata";
-    openDataFile();
+    openDataFile(tif);
+    this->tablePool[tableName] = tif;
+    return tif;
 }
 
 void RecordManager::dropTable(string tableName){
-    if (this->tinfo != nullptr){
-        this->tinfo->writeBack();
+    if (this->tablePool.find(tableName) == this->tablePool.end()){
+        cout << "Table " << tableName << " doesn't exist!\n";
+        return;
     }
-    this->tinfo = nullptr;
-    this->fixedFile = nullptr;
-    this->unfixedFile = nullptr;
+    this->tablePool[tableName] = nullptr;
     unlink(string(tableName+".tbinfo").c_str());
     unlink(string(tableName+".tbdata").c_str());
 }
 
-void RecordManager::openDataFile(){
-    this->fixedFile = nullptr;
-    this->unfixedFile = nullptr;
-    if (this->tinfo->unfixedColNumbers == 0){
-        this->fixedFile = make_shared<FixedRecordFile>(this->dataFileName);
-        this->fixedFile->setRecordLength(this->tinfo->getFixedRecordLength());
-        this->isFixed = true;
-    } else {
-        this->unfixedFile = make_shared<UnfixedRecordFile>(this->dataFileName);
-        this->isFixed = false;
-    }
-    this->tinfo->genConverter();
+void RecordManager::openDataFile(shared_ptr<TableInfo> p){
+    p->openDataFile();
 }
